+++ src/photon/tests/pathTest.php
@@ -34,6 +34,7 @@ class PathTest extends TestCase
     public function testListFiles()
     {
         // review this test, adding a file in the data folder breaks it...
-        $this->assertEquals(13, count(Dir::listFiles(__DIR__ . '/../data')));
+        // Same if we remove a file... We really need to do somethings !
+        $this->assertEquals(2, count(Dir::listFiles(__DIR__ . '/../data')));
     }
 }